# public/api/openapi.yaml
openapi: 3.0.3
info:
  title: InkBoard API
  description: |
    DeFi aggregation dashboard API for the Ink Network ecosystem (chainId: 57073).
    Provides portfolio tracking, yield aggregation, security scanning, cross-chain swap quotes,
    and market data across 5+ Ink protocols.

    ## Authentication
    No API key required for read-only endpoints. Rate limits apply per IP address.

    ## Rate Limits
    - Default: 60 requests/minute per IP
    - Portfolio/DeFi endpoints: 10-20 requests/minute
    - Scan protocol: 5 requests/minute

    ## Data Sources
    - CoinGecko (prices)
    - Velodrome, InkySwap, Tydro, Nado, Curve (yield data)
    - LI.FI (cross-chain swaps)
    - Blockscout (transactions)
  version: 0.2.0
  contact:
    name: Shinka Labs
    url: https://inkboard.pro
    email: contact@shinkalabs.tech
  license:
    name: Proprietary
    url: https://inkboard.pro
servers:
  - url: https://inkboard.pro/api
    description: Production
paths:
...
  /scan-protocol:
    get:
      summary: Scan DeFi protocol for API endpoints
      description: |
        Scans a DeFi protocol frontend for API endpoints, subgraphs, and contract addresses
        relevant to the Ink ecosystem. Only allowlisted DeFi domains are permitted.
      operationId: scanProtocol
      tags:
        - Security
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL of the DeFi protocol to scan (must be an allowlisted domain)
      responses:
        '200':
          description: Scan results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '400':
          description: Invalid or disallowed URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    ScanResult:
      type: object
      properties:
        scanned:
          type: string
        chain:
          type: object
          properties:
            name:
              type: string
            chainId:
              type: integer
            supported:
              type: boolean
            evidence:
              type: array
              items:
                type: string
        summary:
          type: object
          properties:
            bundlesScanned:
              type: integer
            defiEndpoints:
              type: integer
            otherEndpoints:
              type: integer
            contractAddresses:
              type: integer
        defiEndpoints:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        contractAddresses:
          type: array
          items:
            type: object
    Error:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer